refactor(python): split firewall auth tests by owner
