--- crates/runner/mitm-addon/tests/test_firewall_auth.py
@@ -1,3176 +1,0 @@
-"""Tests for firewall auth header resolution and forwarding."""
-
-import asyncio
-import io
-import json
-import time
-import urllib.error
-import uuid
-from email.message import Message
-from unittest.mock import AsyncMock, MagicMock, patch
-
-import pytest
-
-import auth
-import auth_base_forwarder
-import firewall_auth_cache as auth_cache
-import firewall_auth_client as auth_client
-import flow_metadata_keys as metadata_keys
-import matching
-import platform_api
-from aws_sigv4 import AwsSigV4Credentials
-from tests.auth_endpoint_helpers import FakeAuthEndpoint, firewall_auth_success_response
-from tests.auth_state_helpers import (
-    auth_cache_key,
-    cached_headers,
-    force_refresh_pending,
-    last_force_refresh_monotonic_at,
-    mark_force_refresh,
-    require_cached_headers,
-    require_last_force_refresh_monotonic_at,
-    set_cached_headers,
-    set_last_force_refresh_monotonic_at,
-)
-from tests.aws_sigv4_helpers import (
-    DEFAULT_SIGV4_TIMESTAMP,
-    RESOLVED_AWS_SESSION_TOKEN,
-    STS_FORM_BODY,
-    STS_HOST,
-    aws_sigv4_authorization,
-    resolved_aws_sigv4_credentials,
-)
-from tests.firewall_auth_helpers import (
-    apply_requestheaders_auth_without_upstream_admission,
-    handle_firewall_request_without_upstream_admission,
-)
-from tests.firewall_helpers import cancel_pending_task
-from tests.jsonl_log_helpers import read_jsonl_text_after_flush
-from url_utils import get_trusted_authority
-
-_MALFORMED_SUCCESS_PREFIX = "Firewall auth endpoint returned malformed success response"
-_MISSING_FIELD = object()
-
-
-def _fail_if_ordinary_upstream_credentials_are_revalidated() -> bool:
-    raise AssertionError("ordinary upstream credential guard must not run")
-
-
-def _upstream_headers(*pairs: tuple[str, str]) -> Message:
-    headers = Message()
-    for name, value in pairs:
-        headers[name] = value
-    return headers
-
-
-def _http_error(url: str, status: int, reason: str, body: bytes) -> urllib.error.HTTPError:
-    return urllib.error.HTTPError(url, status, reason, Message(), io.BytesIO(body))
-
-
-def _auth_success(
-    *,
-    headers: dict[str, str],
-    expires_at: int | float | None = None,
-    resolved_secrets: list[str] | None = None,
-    refreshed_connectors: list[str] | None = None,
-    refreshed_secrets: list[str] | None = None,
-    base: str | None = None,
-    query: dict[str, str] | None = None,
-    aws_sigv4: AwsSigV4Credentials | None = None,
-) -> auth_client.FirewallAuthSuccess:
-    return auth_client.FirewallAuthSuccess(
-        payload=auth_client.FirewallAuthPayload(
-            headers=headers,
-            resolved_secrets=resolved_secrets or [],
-            base=base,
-            query=query,
-            aws_sigv4=aws_sigv4,
-        ),
-        expires_at=expires_at,
-        refreshed_connectors=refreshed_connectors or [],
-        refreshed_secrets=refreshed_secrets or [],
-    )
-
-
-def _firewall_auth_request(
-    encrypted_secrets: str = "iv:tag:data",
-    auth_headers: dict[str, str] | None = None,
-    sandbox_auth: str = "tok-xyz",
-    **kwargs,
-) -> auth_client.FirewallAuthRequest:
-    return auth_client.FirewallAuthRequest(
-        encrypted_secrets=encrypted_secrets,
-        auth_headers=auth_headers or {},
-        sandbox_token=sandbox_auth,
-        **kwargs,
-    )
-
-
-def _json_response(body: object) -> MagicMock:
-    mock_resp = MagicMock()
-    mock_resp.__enter__.return_value = mock_resp
-    mock_resp.read.return_value = json.dumps(body).encode()
-    return mock_resp
-
-
-def _raw_response(body: bytes) -> MagicMock:
-    mock_resp = MagicMock()
-    mock_resp.__enter__.return_value = mock_resp
-    mock_resp.read.return_value = body
-    return mock_resp
-
-
-def _allow(
-    api_entry: dict,
-    *,
-    name: str = "github",
-    permission: str | None = "repo-read",
-    params: dict[str, str] | None = None,
-    rule: str | None = "GET /repos/{owner}/{repo}",
-    rel_path: str = "/",
-) -> matching.FirewallAllow:
-    return matching.FirewallAllow(api_entry, name, permission, dict(params or {}), rule, rel_path)
-
-
-def _firewall_flow(
-    real_flow,
-    *,
-    host: str = "api.github.com",
-    path: str = "/repos",
-    run_id: str = "test-run",
-):
-    flow = real_flow(with_response=False, host=host, path=path)
-    flow.metadata[metadata_keys.VM_RUN_ID] = run_id
-    return flow
-
-
-def _api_entry(
-    *,
-    base: str = "https://api.github.com",
-    auth_config: dict | None = None,
-    api_id: str | None = None,
-) -> dict:
-    auth = _copy_auth_config(auth_config)
-    entry = {
-        "base": base,
-        "auth": auth,
-    }
-    if api_id is not None:
-        entry["id"] = api_id
-    return entry
-
-
-def _copy_auth_config(auth_config: dict | None) -> dict:
-    if auth_config is None:
-        return {"headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"}}
-
-    copied = dict(auth_config)
-    for key in ("headers", "query"):
-        value = copied.get(key)
-        if isinstance(value, dict):
-            copied[key] = dict(value)
-    return copied
-
-
-def _vm_info(
-    tmp_path=None,
-    *,
-    run_id: str = "run-1",
-    sandbox_marker: str = "tok-xyz",
-    encrypted_secrets: str = "iv:tag:data",
-    include_encrypted_secrets: bool = True,
-    billable_firewalls: list[str] | None = None,
-    include_billable_firewalls: bool = True,
-    network_log_path: str | None = None,
-) -> dict:
-    if network_log_path is None:
-        if tmp_path is None:
-            raise ValueError("tmp_path or network_log_path is required")
-        network_log_path = str(tmp_path / "net.jsonl")
-
-    vm_info: dict[str, object] = {
-        "runId": run_id,
-        "sandboxToken": sandbox_marker,
-        "networkLogPath": network_log_path,
-    }
-    if include_encrypted_secrets:
-        vm_info["encryptedSecrets"] = encrypted_secrets
-    if include_billable_firewalls:
-        vm_info["billableFirewalls"] = list(billable_firewalls or [])
-    return vm_info
-
-
-def _token_meta(
-    *,
-    headers: dict[str, str] | None = None,
-    resolved_secrets: list[str] | None = None,
-    refreshed_connectors: list[str] | None = None,
-    refreshed_secrets: list[str] | None = None,
-    cache_hit: bool = False,
-) -> dict:
-    return {
-        "headers": dict(headers or {}),
-        "resolved_secrets": list(resolved_secrets or []),
-        "refreshed_connectors": list(refreshed_connectors or []),
-        "refreshed_secrets": list(refreshed_secrets or []),
-        "cache_hit": cache_hit,
-    }
-
-
-class _UnreadableHttpErrorBody(io.BytesIO):
-    def read(self, size: int = -1) -> bytes:
-        raise OSError("body read failed")
-
-
-class TestGetFirewallHeaders:
-    async def test_cache_miss_fetches_and_caches(self, headers):
-        mock_headers = {"Authorization": "Bearer fresh-token"}
-        mock_result = _auth_success(headers=mock_headers)
-        encrypted = "iv:tag:data"
-        auth_templates = {"Authorization": "Bearer ${{ secrets.TOKEN }}"}
-        cache_key = auth_cache_key(api_id="https://api.github.com")
-
-        mock_fetch = AsyncMock(return_value=mock_result)
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(
-                    encrypted_secrets=encrypted,
-                    auth_headers=auth_templates,
-                ),
-            )
-
-        assert headers["headers"] == mock_headers
-        assert headers["cache_hit"] is False
-        assert headers["refreshed_connectors"] == []
-        assert headers["refreshed_secrets"] == []
-        mock_fetch.assert_called_once()
-        assert mock_fetch.call_args.args == (
-            _firewall_auth_request(
-                encrypted_secrets=encrypted,
-                auth_headers=auth_templates,
-            ),
-        )
-        assert mock_fetch.call_args.kwargs == {"force_refresh": False}
-
-        assert cached_headers(cache_key)
-        assert require_cached_headers(cache_key).headers == mock_headers
-
-    async def test_cache_hit_returns_cached(self, headers):
-        cache_key = auth_cache_key(api_id="https://api.github.com")
-        cached_headers = {"Authorization": "Bearer cached-token"}
-        set_cached_headers(cache_key, headers=cached_headers)
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert headers["headers"] == cached_headers
-        assert headers["cache_hit"] is True
-        mock_fetch.assert_not_called()
-
-    async def test_cache_hit_with_valid_ttl_returns_cached(self, headers):
-        """Cached entry with expiresAt in the future should be returned without fetching."""
-        cache_key = auth_cache_key()
-        cached_headers = {"Authorization": "Bearer valid-token"}
-        set_cached_headers(
-            cache_key,
-            headers=cached_headers,
-            expires_at=time.time() + 3600,  # 1 hour from now
-        )
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert headers["headers"] == cached_headers
-        assert headers["cache_hit"] is True
-        mock_fetch.assert_not_called()
-
-    async def test_cache_evicted_when_ttl_expired(self, headers):
-        """Cached entry with expiresAt in the past should trigger a re-fetch."""
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer stale-token"},
-            expires_at=time.time() - 10,  # expired 10 seconds ago
-        )
-
-        fresh_headers = {"Authorization": "Bearer fresh-token"}
-        mock_result = _auth_success(headers=fresh_headers, expires_at=time.time() + 3600)
-
-        mock_fetch = AsyncMock(return_value=mock_result)
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert headers["headers"] == fresh_headers
-        assert headers["cache_hit"] is False
-        # fetch_firewall_headers wraps urllib; pins the TTL-expiry→re-fetch contract (#9991).
-        mock_fetch.assert_called_once()
-        # Verify cache was updated with new entry
-        assert require_cached_headers(cache_key).headers == fresh_headers
-
-    async def test_cache_with_null_expires_at_never_evicts(self, headers):
-        """Cached entry with expiresAt=None (non-expiring) should never be evicted by TTL."""
-        cache_key = auth_cache_key()
-        cached_headers = {"Authorization": "Bearer permanent-token"}
-        set_cached_headers(cache_key, headers=cached_headers, expires_at=None)
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert headers["headers"] == cached_headers
-        assert headers["cache_hit"] is True
-        mock_fetch.assert_not_called()
-
-    async def test_billable_cache_hit_requires_valid_expiry(self, headers):
-        cache_key = auth_cache_key()
-        cached_headers = {"Authorization": "Bearer cached-token"}
-        set_cached_headers(cache_key, headers=cached_headers, expires_at=time.time() + 30)
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(firewall_billable=True),
-            )
-
-        assert headers["headers"] == cached_headers
-        assert headers["cache_hit"] is True
-        mock_fetch.assert_not_called()
-
-    @pytest.mark.parametrize(
-        ("expiry", "now"),
-        [
-            pytest.param(None, 100.0, id="none"),
-            pytest.param(True, 0.0, id="bool-true"),
-            pytest.param(False, -1.0, id="bool-false"),
-            pytest.param("123", 100.0, id="string"),
-            pytest.param(float("inf"), 100.0, id="infinity"),
-            pytest.param(float("nan"), 100.0, id="nan"),
-            pytest.param(100.0, 100.0, id="exact-now"),
-        ],
-    )
-    def test_expiry_validation_rejects_invalid_values(self, expiry, now):
-        assert auth_cache._has_valid_expiry(expiry, now=now) is False
-
-    @pytest.mark.parametrize("expiry", [True, "123", float("inf"), float("nan")])
-    async def test_cache_with_invalid_expiry_refetches(self, headers, expiry):
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer malformed-token"},
-            expires_at=expiry,
-        )
-        fresh_headers = {"Authorization": "Bearer fresh-token"}
-        mock_fetch = AsyncMock(return_value=_auth_success(headers=fresh_headers, expires_at=None))
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert headers["headers"] == fresh_headers
-        assert headers["cache_hit"] is False
-        mock_fetch.assert_called_once()
-
-    async def test_billable_cache_without_expiry_refetches(self, headers):
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer stale-token"},
-            expires_at=None,
-        )
-        fresh_headers = {"Authorization": "Bearer fresh-token"}
-        expires_at = int(time.time()) + 30
-        mock_fetch = AsyncMock(
-            return_value=_auth_success(headers=fresh_headers, expires_at=expires_at)
-        )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(firewall_billable=True),
-            )
-
-        assert headers["headers"] == fresh_headers
-        assert headers["cache_hit"] is False
-        mock_fetch.assert_called_once()
-        assert mock_fetch.call_args.args[0].firewall_billable is True
-        assert require_cached_headers(cache_key).expires_at == expires_at
-
-    async def test_billable_cache_with_expired_expiry_refetches(self, headers):
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer stale-token"},
-            expires_at=time.time() - 1,
-        )
-        fresh_headers = {"Authorization": "Bearer fresh-token"}
-        mock_fetch = AsyncMock(
-            return_value=_auth_success(headers=fresh_headers, expires_at=time.time() + 30)
-        )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            headers = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(firewall_billable=True),
-            )
-
-        assert headers["headers"] == fresh_headers
-        assert headers["cache_hit"] is False
-        mock_fetch.assert_called_once()
-
-    @pytest.mark.parametrize(
-        "expires_at",
-        [
-            pytest.param(None, id="none"),
-            pytest.param(True, id="bool"),
-            pytest.param("123", id="string"),
-            pytest.param(float("inf"), id="infinity"),
-            pytest.param(float("nan"), id="nan"),
-            pytest.param(0, id="expired"),
-        ],
-    )
-    async def test_billable_fetch_with_invalid_expiry_fails_closed(self, expires_at):
-        cache_key = auth_cache_key()
-        mock_fetch = AsyncMock(
-            return_value=_auth_success(
-                headers={"Authorization": "Bearer token"},
-                expires_at=expires_at,
-            )
-        )
-
-        with (
-            patch.object(auth_cache, "fetch_firewall_headers", mock_fetch),
-            pytest.raises(auth_cache.InvalidBillableAuthExpiryError),
-        ):
-            await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(firewall_billable=True),
-            )
-        assert cached_headers(cache_key) is None
-
-    async def test_cache_hit_includes_base_when_present(self, headers):
-        """Cached entry with 'base' returns it on cache hit."""
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={},
-            resolved_secrets=["WEBHOOK_URL"],
-            base="https://discord.com/api/webhooks/123/abc",
-            expires_at=None,
-        )
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            result = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(auth_base="${{ secrets.WEBHOOK_URL }}"),
-            )
-
-        assert result["base"] == "https://discord.com/api/webhooks/123/abc"
-        assert result["cache_hit"] is True
-        mock_fetch.assert_not_called()
-
-    async def test_query_is_cached_and_returned_on_cache_hit(self):
-        """auth.query is cached after a fetch and returned on cache hit."""
-        cache_key = auth_cache_key()
-        cached_query = {"api_key": "cached-key", "empty_auth": ""}
-        mock_fetch = AsyncMock(
-            return_value=_auth_success(
-                headers={},
-                resolved_secrets=["QUERY_KEY"],
-                query=cached_query,
-            )
-        )
-        request = _firewall_auth_request(
-            auth_query={
-                "api_key": "${{ secrets.QUERY_KEY }}",
-                "empty_auth": "${{ vars.EMPTY }}",
-            }
-        )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            first = await auth_cache.get_firewall_headers(cache_key, request)
-            second = await auth_cache.get_firewall_headers(cache_key, request)
-
-        assert first["query"] == cached_query
-        assert first["cache_hit"] is False
-        assert second["query"] == cached_query
-        assert second["cache_hit"] is True
-        mock_fetch.assert_called_once()
-        assert require_cached_headers(cache_key).query == cached_query
-
-    async def test_base_and_query_are_cached_together(self):
-        """auth.base and auth.query survive the same cache entry."""
-        cache_key = auth_cache_key()
-        cached_base = "https://example.com/webhook/secret"
-        cached_query = {"api_key": "cached-key", "empty_auth": ""}
-        mock_fetch = AsyncMock(
-            return_value=_auth_success(
-                headers={},
-                base=cached_base,
-                query=cached_query,
-            )
-        )
-        request = _firewall_auth_request(
-            auth_base="${{ secrets.WEBHOOK_URL }}",
-            auth_query={
-                "api_key": "${{ secrets.QUERY_KEY }}",
-                "empty_auth": "${{ vars.EMPTY }}",
-            },
-        )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            first = await auth_cache.get_firewall_headers(cache_key, request)
-            second = await auth_cache.get_firewall_headers(cache_key, request)
-
-        assert first["base"] == cached_base
-        assert first["query"] == cached_query
-        assert first["cache_hit"] is False
-        assert second["base"] == cached_base
-        assert second["query"] == cached_query
-        assert second["cache_hit"] is True
-        mock_fetch.assert_called_once()
-        cached = require_cached_headers(cache_key)
-        assert cached.base == cached_base
-        assert cached.query == cached_query
-
-    async def test_cache_hit_omits_base_when_absent(self, headers):
-        """Cached entry without 'base' does not include it in result."""
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer tok"},
-            resolved_secrets=["TOKEN"],
-            expires_at=None,
-        )
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            result = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(
-                    auth_headers={"Authorization": "Bearer ${{ secrets.TOKEN }}"}
-                ),
-            )
-
-        assert "base" not in result
-        assert result["cache_hit"] is True
-
-    async def test_force_refresh_marker_triggers_forced_fetch(self, headers):
-        """When a force-refresh marker is set, the next fetch passes
-        force_refresh=True, the marker is cleared, and the consume timestamp
-        is recorded so the cooldown can suppress re-marking (#9860)."""
-        cache_key = auth_cache_key()
-        mark_force_refresh(cache_key)
-        before = time.monotonic()
-
-        mock_fetch = AsyncMock(return_value=_auth_success(headers={"Authorization": "Bearer new"}))
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        # force_refresh kwarg must be True
-        assert mock_fetch.call_args.kwargs["force_refresh"] is True
-        # Marker cleared after consumption
-        assert not force_refresh_pending(cache_key)
-        # Consume timestamp recorded for cooldown enforcement
-        assert require_last_force_refresh_monotonic_at(cache_key) >= before
-
-    async def test_force_refresh_fetch_failure_still_consumes_marker(self, headers):
-        """A failed forced refresh burns the cooldown and does not cache headers."""
-        cache_key = auth_cache_key()
-        mark_force_refresh(cache_key)
-        before = time.monotonic()
-
-        mock_fetch = AsyncMock(side_effect=ConnectionError("server unreachable"))
-        with (
-            patch.object(auth_cache, "fetch_firewall_headers", mock_fetch),
-            pytest.raises(ConnectionError),
-        ):
-            await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert mock_fetch.call_args.kwargs["force_refresh"] is True
-        assert not force_refresh_pending(cache_key)
-        assert require_last_force_refresh_monotonic_at(cache_key) >= before
-        assert cached_headers(cache_key) is None
-
-    def test_force_refresh_first_owned_state_is_allowed_without_previous_timestamp(self):
-        """An owned auth state has no cooldown timestamp, so the first 401 marks."""
-        cache_key = auth_cache_key()
-        set_cached_headers(cache_key, headers={"Authorization": "Bearer cached-token"})
-
-        with patch.object(auth_cache.time, "monotonic", return_value=10.0):
-            auth_cache.request_force_refresh(cache_key)
-
-        assert force_refresh_pending(cache_key)
-
-    def test_force_refresh_inside_monotonic_cooldown_is_suppressed(self, headers):
-        """Repeated 401s inside the process-local cooldown do not re-mark."""
-        cache_key = auth_cache_key()
-        set_last_force_refresh_monotonic_at(cache_key, 1000.0)
-
-        with patch.object(auth_cache.time, "monotonic", return_value=1119.0):
-            auth_cache.request_force_refresh(cache_key)
-
-        assert not force_refresh_pending(cache_key)
-
-    def test_force_refresh_after_monotonic_cooldown_is_allowed(self, headers):
-        """A 401 after the monotonic cooldown elapsed can re-mark."""
-        cache_key = auth_cache_key()
-        set_last_force_refresh_monotonic_at(cache_key, 1000.0)
-
-        with patch.object(auth_cache.time, "monotonic", return_value=1121.0):
-            auth_cache.request_force_refresh(cache_key)
-
-        assert force_refresh_pending(cache_key)
-
-    async def test_force_refresh_records_monotonic_timestamp_before_fetch_returns(self, headers):
-        """The cooldown is burned before awaiting the forced auth fetch."""
-        cache_key = auth_cache_key()
-        mark_force_refresh(cache_key)
-
-        async def delayed_fetch(*args, **kwargs):
-            assert kwargs["force_refresh"] is True
-            assert require_last_force_refresh_monotonic_at(cache_key) == 1234.0
-            return _auth_success(headers={"Authorization": "Bearer new"})
-
-        with (
-            patch.object(auth_cache.time, "monotonic", return_value=1234.0),
-            patch.object(auth_cache, "fetch_firewall_headers", side_effect=delayed_fetch),
-        ):
-            await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert require_last_force_refresh_monotonic_at(cache_key) == 1234.0
-
-    async def test_non_forced_fetch_does_not_cache_if_marker_appears_in_flight(self, headers):
-        """A 401 marker during a non-forced fetch must win over the cache write."""
-        cache_key = auth_cache_key()
-        fetch_entered = asyncio.Event()
-        allow_fetch_return = asyncio.Event()
-        first_force_refresh_values = []
-
-        async def delayed_fetch(*args, **kwargs):
-            force_refresh = kwargs["force_refresh"]
-            first_force_refresh_values.append(force_refresh)
-            fetch_entered.set()
-            await allow_fetch_return.wait()
-            return _auth_success(
-                headers={"Authorization": "Bearer maybe-stale"},
-                expires_at=time.time() + 3600,
-            )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", side_effect=delayed_fetch):
-            task = asyncio.create_task(
-                auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-            )
-            try:
-                await asyncio.wait_for(fetch_entered.wait(), timeout=5)
-                auth_cache.request_force_refresh(cache_key)
-                allow_fetch_return.set()
-                result = await task
-            finally:
-                allow_fetch_return.set()
-                await cancel_pending_task(task)
-
-        assert first_force_refresh_values == [False]
-        assert result["headers"] == {"Authorization": "Bearer maybe-stale"}
-        assert result["cache_hit"] is False
-        assert cached_headers(cache_key) is None
-        assert force_refresh_pending(cache_key)
-
-        forced_headers = {"Authorization": "Bearer refreshed"}
-        forced_fetch = AsyncMock(
-            return_value=_auth_success(headers=forced_headers, expires_at=time.time() + 3600)
-        )
-        before_forced = time.monotonic()
-
-        with patch.object(auth_cache, "fetch_firewall_headers", forced_fetch):
-            forced_result = await auth_cache.get_firewall_headers(
-                cache_key, _firewall_auth_request()
-            )
-
-        assert forced_fetch.call_args.kwargs["force_refresh"] is True
-        assert forced_result["headers"] == forced_headers
-        assert forced_result["cache_hit"] is False
-        assert not force_refresh_pending(cache_key)
-        assert require_last_force_refresh_monotonic_at(cache_key) >= before_forced
-        assert require_cached_headers(cache_key).headers == forced_headers
-
-    async def test_waiting_request_force_refreshes_after_in_flight_marker(self, headers):
-        """A same-key waiter must not reuse headers from the stale-prone leader fetch."""
-        cache_key = auth_cache_key()
-        first_fetch_entered = asyncio.Event()
-        allow_first_fetch_return = asyncio.Event()
-        force_refresh_values = []
-
-        async def fetch_with_blocked_leader(*args, **kwargs):
-            force_refresh = kwargs["force_refresh"]
-            force_refresh_values.append(force_refresh)
-            if not force_refresh:
-                first_fetch_entered.set()
-                await allow_first_fetch_return.wait()
-                return _auth_success(
-                    headers={"Authorization": "Bearer maybe-stale"},
-                    expires_at=time.time() + 3600,
-                )
-            return _auth_success(
-                headers={"Authorization": "Bearer refreshed"},
-                expires_at=time.time() + 3600,
-            )
-
-        with patch.object(
-            auth_cache, "fetch_firewall_headers", side_effect=fetch_with_blocked_leader
-        ):
-            leader = asyncio.create_task(
-                auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-            )
-            waiter = None
-            try:
-                await asyncio.wait_for(first_fetch_entered.wait(), timeout=5)
-                waiter = asyncio.create_task(
-                    auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-                )
-                auth_cache.request_force_refresh(cache_key)
-                allow_first_fetch_return.set()
-                leader_result, waiter_result = await asyncio.gather(leader, waiter)
-            finally:
-                allow_first_fetch_return.set()
-                for task in (leader, waiter):
-                    await cancel_pending_task(task)
-
-        assert force_refresh_values == [False, True]
-        assert leader_result["headers"] == {"Authorization": "Bearer maybe-stale"}
-        assert waiter_result["headers"] == {"Authorization": "Bearer refreshed"}
-        assert require_cached_headers(cache_key).headers == {"Authorization": "Bearer refreshed"}
-        assert not force_refresh_pending(cache_key)
-
-    async def test_force_refresh_absent_passes_false(self, headers):
-        """Without a marker, fetch is called with force_refresh=False (#9860)."""
-        cache_key = auth_cache_key(api_id="api-2")
-        mock_fetch = AsyncMock(return_value=_auth_success(headers={}))
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert mock_fetch.call_args.kwargs["force_refresh"] is False
-        # No consume timestamp written when force-refresh didn't happen
-        assert last_force_refresh_monotonic_at(cache_key) is None
-
-    async def test_force_refresh_marker_is_scoped_to_auth_identity(self, headers):
-        old_key = auth_cache_key(auth_identity="old-auth")
-        new_key = auth_cache_key(auth_identity="new-auth")
-        mark_force_refresh(old_key)
-
-        mock_fetch = AsyncMock(return_value=_auth_success(headers={}))
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            await auth_cache.get_firewall_headers(new_key, _firewall_auth_request())
-
-        assert mock_fetch.call_args.kwargs["force_refresh"] is False
-        assert force_refresh_pending(old_key)
-        assert not force_refresh_pending(new_key)
-
-    async def test_force_refresh_marker_ignored_on_cache_hit(self, headers):
-        """Fast-path cache hit does NOT consume the force-refresh marker —
-        marker survives until the next actual fetch (#9860)."""
-        cache_key = auth_cache_key()
-        set_cached_headers(
-            cache_key,
-            headers={"Authorization": "Bearer cached"},
-            expires_at=None,
-        )
-        mark_force_refresh(cache_key)
-
-        mock_fetch = AsyncMock()
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch):
-            result = await auth_cache.get_firewall_headers(cache_key, _firewall_auth_request())
-
-        assert result["cache_hit"] is True
-        mock_fetch.assert_not_called()
-        # Marker preserved for next real fetch
-        assert force_refresh_pending(cache_key)
-
-
-# =========================================================================
-# handle_firewall_request
-# =========================================================================
-
-
-class TestHandleFirewallRequest:
-    async def test_success_injects_headers_and_audit_metadata(
-        self, real_flow, headers, mitm_ctx, tmp_path
-    ):
-        flow = _firewall_flow(real_flow)
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow.metadata[metadata_keys.VM_PROXY_LOG_PATH] = str(proxy_log_path)
-        api_entry = _api_entry(
-            api_id="run-1:0",
-            auth_config={"headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"}},
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry, params={"owner": "octocat", "repo": "hello"})
-        token_meta = _token_meta(
-            headers={"Authorization": "Bearer real-token", "X-Custom": "value"},
-            resolved_secrets=["GITHUB_TOKEN"],
-        )
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-
-        # Headers injected
-        assert flow.request.headers["Authorization"] == "Bearer real-token"
-        assert flow.request.headers["X-Custom"] == "value"
-
-        # Token replacement metadata
-        assert flow.metadata[metadata_keys.AUTH_RESOLVED_SECRETS] == ["GITHUB_TOKEN"]
-        assert flow.metadata[metadata_keys.AUTH_REFRESHED_CONNECTORS] == []
-        assert flow.metadata[metadata_keys.AUTH_REFRESHED_SECRETS] == []
-        assert flow.metadata[metadata_keys.AUTH_CACHE_HIT] is False
-
-        # Core metadata
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_BASE] == "https://api.github.com"
-        assert flow.metadata[metadata_keys.FIREWALL_API_ID] == "run-1:0"
-
-        # Audit metadata
-        assert flow.metadata[metadata_keys.FIREWALL_NAME] == "github"
-        assert flow.metadata[metadata_keys.FIREWALL_PERMISSION] == "repo-read"
-        assert flow.metadata[metadata_keys.FIREWALL_RULE_MATCH] == "GET /repos/{owner}/{repo}"
-        assert flow.metadata[metadata_keys.FIREWALL_PARAMS] == {"owner": "octocat", "repo": "hello"}
-        log_text = await asyncio.to_thread(read_jsonl_text_after_flush, proxy_log_path)
-        assert "Firewall https://api.github.com: api.github.com" in log_text
-
-    async def test_empty_auth_config_preserves_existing_authorization(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        flow = real_flow(
-            with_response=False,
-            host="api.cloudflare.com",
-            path="/client/v4/pages/assets/check-missing",
-            method="POST",
-        )
-        flow.request.headers["Authorization"] = "Bearer upload-jwt"
-        flow.metadata[metadata_keys.VM_RUN_ID] = "test-run"
-        api_entry = _api_entry(
-            base="https://api.cloudflare.com/client",
-            auth_config={},
-            api_id="run-1:1",
-        )
-        vm_info = _vm_info(tmp_path, include_encrypted_secrets=False)
-        allow = _allow(
-            api_entry,
-            name="cloudflare",
-            permission="page.write",
-            rule="POST /v4/pages/assets/check-missing",
-            rel_path="/v4/pages/assets/check-missing",
-        )
-        mock_get_firewall_headers = AsyncMock()
-
-        with (
-            patch.object(auth, "get_firewall_headers", mock_get_firewall_headers),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        mock_get_firewall_headers.assert_not_called()
-        assert flow.request.headers["Authorization"] == "Bearer upload-jwt"
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_NAME] == "cloudflare"
-        assert flow.metadata[metadata_keys.FIREWALL_PERMISSION] == "page.write"
-        assert flow.metadata[metadata_keys.AUTH_CACHE_HIT] is False
-
-    async def test_empty_auth_config_requestheaders_preserves_authorization_without_auth_resolution(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        flow = real_flow(
-            with_response=False,
-            host="api.cloudflare.com",
-            path="/client/v4/pages/assets/upload",
-            method="POST",
-        )
-        flow.request.headers["Authorization"] = "Bearer upload-jwt"
-        flow.metadata[metadata_keys.VM_RUN_ID] = "test-run"
-        api_entry = _api_entry(
-            base="https://api.cloudflare.com/client",
-            auth_config={},
-            api_id="run-1:1",
-        )
-        vm_info = _vm_info(tmp_path, include_encrypted_secrets=False)
-        allow = _allow(
-            api_entry,
-            name="cloudflare",
-            permission="page.write",
-            rule="POST /v4/pages/assets/upload",
-            rel_path="/v4/pages/assets/upload",
-        )
-        mock_get_firewall_headers = AsyncMock()
-
-        with (
-            patch.object(auth, "get_firewall_headers", mock_get_firewall_headers),
-            mitm_ctx(),
-        ):
-            result = await apply_requestheaders_auth_without_upstream_admission(
-                flow,
-                allow,
-                vm_info,
-            )
-
-        assert result is auth.FirewallHeaderPhaseAuthResult.APPLIED
-        mock_get_firewall_headers.assert_not_called()
-        assert flow.request.headers["Authorization"] == "Bearer upload-jwt"
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_NAME] == "cloudflare"
-        assert flow.metadata[metadata_keys.FIREWALL_PERMISSION] == "page.write"
-        assert flow.metadata[metadata_keys.AUTH_CACHE_HIT] is False
-
-    async def test_billable_only_auth_does_not_require_upstream_admission(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry(auth_config={})
-        vm_info = _vm_info(tmp_path, billable_firewalls=["github"])
-        allow = _allow(api_entry)
-        token_meta = _token_meta(headers={})
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            result = await auth.handle_firewall_request(
-                flow,
-                allow,
-                vm_info,
-                revalidate_ordinary_upstream_credentials=(
-                    _fail_if_ordinary_upstream_credentials_are_revalidated
-                ),
-            )
-
-        assert result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        assert flow.response is None
-
-    async def test_auth_cache_identity_tracks_request_auth_inputs(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        """Same run/api entries must not share headers across auth input changes."""
-        api_entry = _api_entry(
-            api_id="run-1:0",
-            auth_config={"headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"}},
-        )
-        allow = _allow(api_entry)
-        first_flow = _firewall_flow(real_flow, run_id="run-1")
-        second_flow = _firewall_flow(real_flow, run_id="run-1")
-        first_vm_info = _vm_info(tmp_path, run_id="run-1", encrypted_secrets="encrypted-first")
-        second_vm_info = _vm_info(tmp_path, run_id="run-1", encrypted_secrets="encrypted-second")
-        mock_fetch = AsyncMock(
-            side_effect=[
-                _auth_success(headers={"Authorization": "Bearer first"}),
-                _auth_success(headers={"Authorization": "Bearer second"}),
-            ]
-        )
-
-        with patch.object(auth_cache, "fetch_firewall_headers", mock_fetch), mitm_ctx():
-            first_result = await handle_firewall_request_without_upstream_admission(
-                first_flow, allow, first_vm_info
-            )
-            second_result = await handle_firewall_request_without_upstream_admission(
-                second_flow, allow, second_vm_info
-            )
-
-        assert first_result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        assert second_result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        assert first_flow.request.headers["Authorization"] == "Bearer first"
-        assert second_flow.request.headers["Authorization"] == "Bearer second"
-        assert mock_fetch.call_count == 2
-
-        first_cache_key = first_flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]
-        second_cache_key = second_flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]
-        assert first_cache_key.run_id == "run-1"
-        assert first_cache_key.api_id == "run-1:0"
-        assert second_cache_key.run_id == "run-1"
-        assert second_cache_key.api_id == "run-1:0"
-        assert first_cache_key.auth_identity != second_cache_key.auth_identity
-
-    @pytest.mark.parametrize(
-        "changed_input",
-        [
-            "firewall-name",
-            "firewall-base",
-            "auth-headers",
-            "auth-base",
-            "auth-query",
-            "auth-aws-sigv4",
-            "encrypted-secrets",
-            "secret-connector-map",
-            "secret-connector-metadata-map",
-            "vars",
-            "billable",
-            "sandbox-token",
-        ],
-    )
-    async def test_auth_cache_identity_distinguishes_all_auth_inputs(
-        self, real_flow, mitm_ctx, tmp_path, changed_input
-    ):
-        default_auth = {"headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"}}
-        changed_auth = default_auth
-        first_name = "github"
-        second_name = first_name
-        first_base = "https://api.github.com"
-        second_base = first_base
-        first_vm_info = _vm_info(tmp_path, run_id="run-1")
-        second_vm_info = _vm_info(tmp_path, run_id="run-1")
-
-        if changed_input == "firewall-name":
-            second_name = "github-enterprise"
-        elif changed_input == "firewall-base":
-            second_base = "https://api.github.com/v1"
-        elif changed_input == "auth-headers":
-            changed_auth = {"headers": {"Authorization": "Bearer ${{ secrets.OTHER_TOKEN }}"}}
-        elif changed_input == "auth-base":
-            changed_auth = {
-                **default_auth,
-                "base": "${{ secrets.WEBHOOK_URL }}",
-            }
-        elif changed_input == "auth-query":
-            changed_auth = {
-                **default_auth,
-                "query": {"api_key": "${{ secrets.QUERY_KEY }}"},
-            }
-        elif changed_input == "auth-aws-sigv4":
-            changed_auth = {
-                **default_auth,
-                "awsSigv4": {
-                    "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                    "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                },
-            }
-        elif changed_input == "encrypted-secrets":
-            second_vm_info["encryptedSecrets"] = "encrypted-updated"
-        elif changed_input == "secret-connector-map":
-            second_vm_info["secretConnectorMap"] = {"GITHUB_TOKEN": "github"}
-        elif changed_input == "secret-connector-metadata-map":
-            second_vm_info["secretConnectorMetadataMap"] = {"GITHUB_TOKEN": {"kind": "oauth"}}
-        elif changed_input == "vars":
-            second_vm_info["vars"] = {"TEAM": "vm0"}
-        elif changed_input == "billable":
-            second_vm_info["billableFirewalls"] = [second_name]
-        elif changed_input == "sandbox-token":
-            second_vm_info["sandboxToken"] = "sandbox-token-updated"
-        else:
-            raise AssertionError(f"Unhandled auth identity input: {changed_input}")
-
-        first_api_entry = _api_entry(
-            base=first_base,
-            api_id="run-1:0",
-            auth_config=default_auth,
-        )
-        second_api_entry = _api_entry(
-            base=second_base,
-            api_id="run-1:0",
-            auth_config=changed_auth,
-        )
-        first_flow = _firewall_flow(real_flow, run_id="run-1")
-        second_flow = _firewall_flow(real_flow, run_id="run-1")
-        mock_get_firewall_headers = AsyncMock(return_value=_token_meta())
-
-        with (
-            patch.object(auth, "get_firewall_headers", mock_get_firewall_headers),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(
-                first_flow,
-                _allow(first_api_entry, name=first_name),
-                first_vm_info,
-            )
-            await handle_firewall_request_without_upstream_admission(
-                second_flow,
-                _allow(second_api_entry, name=second_name),
-                second_vm_info,
-            )
-
-        assert mock_get_firewall_headers.await_count == 2
-        first_cache_key = first_flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]
-        second_cache_key = second_flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]
-        assert first_cache_key.run_id == second_cache_key.run_id == "run-1"
-        assert first_cache_key.api_id == second_cache_key.api_id == "run-1:0"
-        assert first_cache_key.auth_identity != second_cache_key.auth_identity
-
-    async def test_standard_auth_filters_unsafe_resolved_headers(
-        self, real_flow, headers, mitm_ctx, tmp_path
-    ):
-        flow = _firewall_flow(
-            real_flow,
-            path="/repos?existing=1",
-        )
-        api_entry = _api_entry(
-            auth_config={
-                "headers": dict.fromkeys(
-                    (
-                        "Connection",
-                        "Host",
-                        "Content-Length",
-                        "Transfer-Encoding",
-                        "Keep-Alive",
-                        "Proxy-Authenticate",
-                        "Proxy-Authorization",
-                        "Proxy-Connection",
-                        "TE",
-                        "Trailer",
-                        "Upgrade",
-                        "Authorization",
-                        "X-Injected",
-                    ),
-                    "${{ secrets.VALUE }}",
-                ),
-                "query": {"api_key": "${{ secrets.API_KEY }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        token_meta = _token_meta(
-            headers={
-                "Connection": "Authorization, X-Injected",
-                "Host": "evil.example.com",
-                "Content-Length": "999",
-                "Transfer-Encoding": "chunked",
-                "Keep-Alive": "timeout=5",
-                "Proxy-Authenticate": "Basic realm=proxy",
-                "Proxy-Authorization": "Basic secret",
-                "Proxy-Connection": "keep-alive",
-                "TE": "trailers",
-                "Trailer": "X-Trailer",
-                "Upgrade": "websocket",
-                "Authorization": "Bearer real-token",
-                "X-Injected": "trusted",
-            },
-        )
-        token_meta["query"] = {"api_key": "secret"}
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        header_names = {name.lower() for name, _value in flow.request.headers.items(multi=True)}
-        assert flow.request.headers["Host"] == "api.github.com"
-        assert "connection" not in header_names
-        assert "content-length" not in header_names
-        assert "transfer-encoding" not in header_names
-        assert "keep-alive" not in header_names
-        assert "proxy-authenticate" not in header_names
-        assert "proxy-authorization" not in header_names
-        assert "proxy-connection" not in header_names
-        assert "te" not in header_names
-        assert "trailer" not in header_names
-        assert "upgrade" not in header_names
-        assert "authorization" not in header_names
-        assert "x-injected" not in header_names
-        assert flow.request.query["existing"] == "1"
-        assert flow.request.query["api_key"] == "secret"
-
-    async def test_standard_auth_filters_unsafe_headers_before_aws_sigv4_signing(
-        self, headers, real_flow, mitm_ctx, tmp_path
-    ):
-        placeholder_authorization = aws_sigv4_authorization(
-            signed_headers="content-type;host;x-amz-date"
-        )
-        flow = real_flow(
-            with_response=False,
-            host=STS_HOST,
-            path="/",
-            method="POST",
-            request_body=STS_FORM_BODY,
-            request_headers=headers(
-                ("Host", STS_HOST),
-                ("Content-Type", "application/x-www-form-urlencoded"),
-                ("X-Amz-Date", DEFAULT_SIGV4_TIMESTAMP),
-                ("Authorization", placeholder_authorization),
-            ),
-        )
-        flow.metadata[metadata_keys.VM_RUN_ID] = "test-run"
-        flow.metadata[metadata_keys.ORIGINAL_URL] = get_trusted_authority(flow).url
-        api_entry = _api_entry(
-            base="https://sts.amazonaws.com",
-            auth_config={
-                "headers": {"Host": "${{ secrets.UNSAFE_HOST }}"},
-                "awsSigv4": {
-                    "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                    "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                    "sessionToken": "${{ secrets.AWS_SESSION_TOKEN }}",
-                },
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry, rule="POST /", rel_path="/")
-        token_meta = _token_meta(
-            headers={
-                "Host": "evil.example.com",
-                "X-Amz-Meta-Test": "trusted-meta",
-            },
-        )
-        token_meta["aws_sigv4"] = resolved_aws_sigv4_credentials()
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.CONTINUE_UPSTREAM
-        assert flow.request.headers["host"] == STS_HOST
-        assert flow.request.headers["x-amz-meta-test"] == "trusted-meta"
-        assert flow.request.headers["x-amz-security-token"] == RESOLVED_AWS_SESSION_TOKEN
-        assert (
-            "Credential=AKIDEXAMPLE/20260101/us-east-1/sts/aws4_request"
-            in flow.request.headers["authorization"]
-        )
-        assert "evil.example.com" not in flow.request.headers["authorization"]
-
-    @pytest.mark.parametrize(
-        "resolved_headers",
-        [
-            pytest.param({"": "value"}, id="empty-name"),
-            pytest.param({"Bad\nName": "value"}, id="newline-name"),
-            pytest.param({":authority": "evil.example.com"}, id="pseudo-header-name"),
-            pytest.param({"X-Test": "bad\r\nX-Injected: value"}, id="newline-value"),
-        ],
-    )
-    async def test_standard_auth_rejects_malformed_resolved_headers(
-        self,
-        resolved_headers: dict[str, str],
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        token_meta = _token_meta(
-            headers={
-                **resolved_headers,
-                "Authorization": "Bearer real-token",
-            },
-        )
-        token_meta["query"] = {"api_key": "secret"}
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "invalid_resolved_auth_header"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "invalid_resolved_auth_header"
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-
-    async def test_missing_billable_firewalls_falls_back_to_empty(
-        self, real_flow, headers, mitm_ctx, tmp_path
-    ):
-        """billableFirewalls is optional in the TS schema — a vm_info without
-        the key must not KeyError; firewall_billable should be False."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry(api_id="run-1:0")
-        vm_info = _vm_info(tmp_path, include_billable_firewalls=False)
-        allow = _allow(api_entry, rule="GET /repos")
-        token_meta = _token_meta()
-
-        with (
-            patch.object(auth, "get_firewall_headers", AsyncMock(return_value=token_meta)),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.metadata[metadata_keys.FIREWALL_BILLABLE] is False
-
-    async def test_failure_returns_502(self, real_flow, headers, mitm_ctx, tmp_path):
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(side_effect=Exception("API unreachable")),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert "API unreachable" in body["message"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-
-    @pytest.mark.parametrize(
-        ("network_error", "expected_message"),
-        [
-            (
-                urllib.error.URLError("connection refused"),
-                "connection refused",
-            ),
-            (
-                TimeoutError("timed out"),
-                "timed out",
-            ),
-            (
-                ConnectionResetError("connection reset"),
-                "connection reset",
-            ),
-        ],
-        ids=["url-error", "socket-timeout", "connection-reset"],
-    )
-    async def test_auth_endpoint_transport_failure_returns_502(
-        self,
-        network_error: Exception,
-        expected_message: str,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-
-        with (
-            patch("firewall_auth_client._opener.open", side_effect=network_error),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert expected_message in body["message"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-
-    async def test_malformed_success_response_returns_502_without_auth_mutation(
-        self,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        mock_resp = MagicMock()
-        mock_resp.__enter__.return_value = mock_resp
-        mock_resp.__exit__.return_value = False
-        mock_resp.read.return_value = b"not-json"
-
-        with (
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-        mock_resp.__exit__.assert_called_once()
-
-    @pytest.mark.parametrize(
-        ("field_name", "invalid_value", "expected_reason"),
-        [
-            pytest.param(
-                "expiresAt",
-                _MISSING_FIELD,
-                "expiresAt is required",
-                id="expires-at-missing",
-            ),
-            pytest.param(
-                "expiresAt",
-                True,
-                "expiresAt must be a finite number or null",
-                id="expires-at-bool",
-            ),
-            pytest.param(
-                "expiresAt",
-                "123",
-                "expiresAt must be a finite number or null",
-                id="expires-at-string",
-            ),
-            pytest.param(
-                "expiresAt",
-                [],
-                "expiresAt must be a finite number or null",
-                id="expires-at-array",
-            ),
-            pytest.param(
-                "expiresAt",
-                float("nan"),
-                "expiresAt must be a finite number or null",
-                id="expires-at-nan",
-            ),
-            pytest.param(
-                "expiresAt",
-                float("inf"),
-                "expiresAt must be a finite number or null",
-                id="expires-at-positive-infinity",
-            ),
-            pytest.param(
-                "expiresAt",
-                float("-inf"),
-                "expiresAt must be a finite number or null",
-                id="expires-at-negative-infinity",
-            ),
-            pytest.param(
-                "resolvedSecrets",
-                _MISSING_FIELD,
-                "resolvedSecrets is required",
-                id="resolved-secrets-missing",
-            ),
-            pytest.param(
-                "resolvedSecrets",
-                None,
-                "resolvedSecrets must be an array",
-                id="resolved-secrets-null",
-            ),
-            pytest.param(
-                "resolvedSecrets",
-                "TOKEN",
-                "resolvedSecrets must be an array",
-                id="resolved-secrets-string",
-            ),
-            pytest.param(
-                "resolvedSecrets",
-                [123],
-                "resolvedSecrets values must be strings",
-                id="resolved-secrets-item-number",
-            ),
-            pytest.param(
-                "refreshedConnectors",
-                _MISSING_FIELD,
-                "refreshedConnectors is required",
-                id="refreshed-connectors-missing",
-            ),
-            pytest.param(
-                "refreshedConnectors",
-                None,
-                "refreshedConnectors must be an array",
-                id="refreshed-connectors-null",
-            ),
-            pytest.param(
-                "refreshedConnectors",
-                "github",
-                "refreshedConnectors must be an array",
-                id="refreshed-connectors-string",
-            ),
-            pytest.param(
-                "refreshedConnectors",
-                [123],
-                "refreshedConnectors values must be strings",
-                id="refreshed-connectors-item-number",
-            ),
-            pytest.param(
-                "refreshedSecrets",
-                _MISSING_FIELD,
-                "refreshedSecrets is required",
-                id="refreshed-secrets-missing",
-            ),
-            pytest.param(
-                "refreshedSecrets",
-                None,
-                "refreshedSecrets must be an array",
-                id="refreshed-secrets-null",
-            ),
-            pytest.param(
-                "refreshedSecrets",
-                "TOKEN",
-                "refreshedSecrets must be an array",
-                id="refreshed-secrets-string",
-            ),
-            pytest.param(
-                "refreshedSecrets",
-                [None],
-                "refreshedSecrets values must be strings",
-                id="refreshed-secrets-item-null",
-            ),
-        ],
-    )
-    async def test_invalid_required_success_metadata_is_not_cached_or_applied(
-        self,
-        field_name: str,
-        invalid_value: object,
-        expected_reason: str,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        response = firewall_auth_success_response(
-            {"Authorization": "Bearer resolved"},
-            resolved_secrets=["GITHUB_TOKEN"],
-        )
-        response["query"] = {"api_key": "resolved-key"}
-        if invalid_value is _MISSING_FIELD:
-            response.pop(field_name)
-        else:
-            response[field_name] = invalid_value
-        mock_resp = _json_response(response)
-
-        with (
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(
-                flow,
-                allow,
-                vm_info,
-            )
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert body["message"] == (
-            f"Failed to resolve auth headers: {_MALFORMED_SUCCESS_PREFIX}: {expected_reason}"
-        )
-        mock_resp.__exit__.assert_called_once()
-
-    async def test_strategy_inconsistent_success_returns_502_without_auth_mutation(
-        self,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "base": "${{ secrets.WEBHOOK_URL }}",
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        mock_resp = _json_response(
-            firewall_auth_success_response({"Authorization": "Bearer resolved"})
-            | {
-                "query": {"api_key": "resolved-key"},
-            }
-        )
-
-        with (
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert json.loads(flow.response.content)["error"] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        assert metadata_keys.AUTH_URL_REWRITE not in flow.metadata
-
-    async def test_oversized_success_response_returns_502_without_auth_mutation(
-        self,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        response_body = json.dumps({"headers": {"Authorization": "Bearer tok"}}).encode()
-        mock_resp = _raw_response(response_body)
-
-        with (
-            patch.object(
-                auth_client,
-                "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES",
-                len(response_body) - 1,
-            ),
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert "Firewall auth response body too large" in body["message"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-        mock_resp.__exit__.assert_called_once()
-
-    async def test_malformed_json_success_response_returns_502_without_auth_mutation(
-        self,
-        real_flow,
-        mitm_ctx,
-        tmp_path,
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        mock_resp = _json_response({"headers": []})
-
-        with (
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_failed"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_failed"
-        assert _MALFORMED_SUCCESS_PREFIX in body["message"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-        mock_resp.__exit__.assert_called_once()
-
-    async def test_structured_api_error_is_preserved(self, real_flow, mitm_ctx, tmp_path):
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        api_error = auth_client.FirewallAuthApiError(
-            status=502,
-            code="TOKEN_REFRESH_FAILED",
-            message="Access token expired and refresh failed for: codex-oauth-token.",
-            connectors=["codex-oauth-token"],
-            failure_reason="upstream_provider",
-        )
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(side_effect=api_error),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "TOKEN_REFRESH_FAILED"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "TOKEN_REFRESH_FAILED"
-        assert body["message"] == "Access token expired and refresh failed for: codex-oauth-token."
-        assert body["permission"] == "github"
-        assert body["connectors"] == ["codex-oauth-token"]
-        assert body["failureReason"] == "upstream_provider"
-
-    async def test_structured_api_4xx_error_blocks_without_auth_mutation(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        flow = _firewall_flow(real_flow, path="/repos?existing=1")
-        api_entry = _api_entry(
-            auth_config={
-                "headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"},
-                "query": {"api_key": "${{ secrets.GITHUB_TOKEN }}"},
-            },
-        )
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        api_error = auth_client.FirewallAuthApiError(
-            status=403,
-            code="FORBIDDEN",
-            message="Firewall auth denied",
-        )
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(side_effect=api_error),
-            ),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 403
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "FORBIDDEN"
-        assert "Authorization" not in flow.request.headers
-        assert "api_key" not in flow.request.query
-        assert flow.request.query["existing"] == "1"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "FORBIDDEN"
-        assert body["message"] == "Firewall auth denied"
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-
-    async def test_invalid_billable_auth_expiry_returns_502(self, real_flow, mitm_ctx, tmp_path):
-        flow = _firewall_flow(real_flow)
-        proxy_log_path = tmp_path / "proxy.jsonl"
-        flow.metadata[metadata_keys.VM_PROXY_LOG_PATH] = str(proxy_log_path)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path, billable_firewalls=["github"])
-        allow = _allow(api_entry)
-
-        with (
-            patch.object(
-                auth_cache,
-                "fetch_firewall_headers",
-                AsyncMock(
-                    return_value=_auth_success(
-                        headers={"Authorization": "Bearer token"},
-                        base="https://forward.example/secret",
-                        query={"api_key": "secret"},
-                        expires_at=None,
-                    )
-                ),
-            ),
-            mitm_ctx(),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "invalid_auth_expiry"
-        assert "Authorization" not in flow.request.headers
-        assert metadata_keys.AUTH_URL_REWRITE not in flow.metadata
-        assert "api_key" not in flow.request.query
-        assert cached_headers(flow.metadata[metadata_keys.FIREWALL_AUTH_CACHE_KEY]) is None
-        body = json.loads(flow.response.content)
-        assert body["error"] == "invalid_auth_expiry"
-        assert "valid cache expiry" in body["message"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-        log_text = await asyncio.to_thread(read_jsonl_text_after_flush, proxy_log_path)
-        assert "invalid expiresAt" in log_text
-
-    async def test_no_response_set_on_success(self, real_flow, headers, mitm_ctx):
-        """On success, flow.response should remain None (request continues to origin)."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(network_log_path="")
-        allow = _allow(api_entry)
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(
-                    return_value={
-                        "headers": {"Auth": "tok"},
-                        "resolved_secrets": [],
-                        "refreshed_connectors": [],
-                        "refreshed_secrets": [],
-                        "cache_hit": False,
-                    }
-                ),
-            ),
-            mitm_ctx(),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is None
-
-    async def test_connector_not_configured_returns_424(
-        self, real_flow, headers, mitm_ctx, tmp_path
-    ):
-        """When connector is enabled but not linked, return 424 with missing secrets."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(
-                    side_effect=auth_client.ConnectorNotConfiguredError(
-                        "Connector not configured",
-                    )
-                ),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 424
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "connector_not_configured"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "connector_not_configured"
-        assert body["message"] == "Connector not configured"
-        assert body["connectors"] == ["github"]
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-
-    async def test_insufficient_credits_returns_402(self, real_flow, headers, mitm_ctx, tmp_path):
-        """Billable firewall auth denied for credits returns 402 and blocks usage."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path, billable_firewalls=["github"])
-        allow = _allow(api_entry)
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(side_effect=auth_client.InsufficientCreditsError("Insufficient credits")),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 402
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "insufficient_credits"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "insufficient_credits"
-        assert body["message"] == "Insufficient credits"
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-
-    async def test_connector_not_configured_without_name_omits_connectors(
-        self, real_flow, headers, mitm_ctx, tmp_path
-    ):
-        """Connector slugs are only returned when the firewall name is known."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry, name="", permission=None, rule=None)
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(
-                    side_effect=auth_client.ConnectorNotConfiguredError(
-                        "Connector not configured",
-                    )
-                ),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 424
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "BLOCK"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "connector_not_configured"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "connector_not_configured"
-        assert body["message"] == "Connector not configured"
-        assert body["permission"] == ""
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-
-    async def test_missing_vars_only_returns_424(self, real_flow, headers, mitm_ctx, tmp_path):
-        """When connector is not configured, return 424 with its slug."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry(base="https://hcti.io")
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry, name="htmlcsstoimage")
-
-        with (
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(
-                    side_effect=auth_client.ConnectorNotConfiguredError(
-                        "Connector not configured",
-                    )
-                ),
-            ),
-            mitm_ctx(),
-            patch.object(platform_api, "get_api_url", return_value="https://api.vm0.ai"),
-        ):
-            await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert flow.response is not None
-        assert flow.response.status_code == 424
-        body = json.loads(flow.response.content)
-        assert body["error"] == "connector_not_configured"
-        assert body["connectors"] == ["htmlcsstoimage"]
-        assert body["base"] == "https://hcti.io"
-
-    async def test_missing_encrypted_secrets_returns_502(self, real_flow, headers, mitm_ctx):
-        """When encryptedSecrets is missing from vm_info, return 502."""
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry(
-            auth_config={"headers": {"Authorization": "Bearer ${{ secrets.GITHUB_TOKEN }}"}}
-        )
-        vm_info = _vm_info(network_log_path="", include_encrypted_secrets=False)
-        allow = _allow(api_entry)
-        admission = auth_base_forwarder.reserve_forward_request_admission(42)
-        auth_base_forwarder.attach_forward_request_admission_to_flow(flow, admission)
-
-        with mitm_ctx():
-            result = await handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-
-        assert result is auth.FirewallAuthHandlingResult.LOCAL_RESPONSE
-        assert flow.response is not None
-        assert flow.response.status_code == 502
-        assert flow.metadata[metadata_keys.FIREWALL_ACTION] == "ALLOW"
-        assert flow.metadata[metadata_keys.FIREWALL_ERROR] == "auth_unavailable"
-        body = json.loads(flow.response.content)
-        assert body["error"] == "auth_unavailable"
-        assert body["message"] == "Auth secrets not configured"
-        assert body["permission"] == "github"
-        assert body["base"] == "https://api.github.com"
-        assert "connectors" not in body
-        assert metadata_keys.AUTH_BASE_FORWARD_ADMISSION not in flow.metadata
-        assert auth_base_forwarder.forward_request_admission_state_for_tests() == (0, 0)
-
-    async def test_cancelled_auth_resolution_releases_forward_admission(
-        self, real_flow, mitm_ctx, tmp_path
-    ):
-        flow = _firewall_flow(real_flow)
-        api_entry = _api_entry()
-        vm_info = _vm_info(tmp_path)
-        allow = _allow(api_entry)
-        admission = auth_base_forwarder.reserve_forward_request_admission(42)
-        auth_base_forwarder.attach_forward_request_admission_to_flow(flow, admission)
-        auth_resolution_entered = asyncio.Event()
-        release_auth_resolution = asyncio.Event()
-
-        async def wait_for_auth_resolution(*_args, **_kwargs):
-            auth_resolution_entered.set()
-            await release_auth_resolution.wait()
-
-        with (
-            mitm_ctx(),
-            patch.object(
-                auth,
-                "get_firewall_headers",
-                AsyncMock(side_effect=wait_for_auth_resolution),
-            ),
-        ):
-            task = asyncio.create_task(
-                handle_firewall_request_without_upstream_admission(flow, allow, vm_info)
-            )
-            try:
-                await asyncio.wait_for(auth_resolution_entered.wait(), timeout=1)
-                task.cancel()
-                with pytest.raises(asyncio.CancelledError):
-                    _ = await task
-            finally:
-                release_auth_resolution.set()
-                await cancel_pending_task(task)
-
-        assert metadata_keys.AUTH_BASE_FORWARD_ADMISSION not in flow.metadata
-        assert auth_base_forwarder.forward_request_admission_state_for_tests() == (0, 0)
-
-
-# =========================================================================
-# fetch_firewall_headers
-# =========================================================================
-
-
-class TestMakeApiRequest:
-    def test_builds_platform_api_request_with_standard_headers(self, mitm_ctx):
-        with (
-            mitm_ctx(
-                client_session_id="runner-session-direct",
-                client_version="runner-version-direct",
-            ),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            req = platform_api.make_api_request(
-                "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-                b"{}",
-                "tok-xyz",
-            )
-
-        assert req.full_url == "https://api.vm0.ai/api/webhooks/agent/firewall/auth"
-        assert req.data == b"{}"
-        headers = dict(req.header_items())
-        assert headers["Content-type"] == "application/json"
-        assert headers["Authorization"] == "Bearer tok-xyz"
-        assert headers["User-agent"] == "vm0-mitm-addon/1.0"
-        normalized_headers = {name.lower(): value for name, value in headers.items()}
-        assert normalized_headers["x-client-version"] == "runner-version-direct"
-        assert normalized_headers["x-client-type"] == "MitmAddon"
-        assert normalized_headers["x-client-session-id"] == "runner-session-direct"
-        uuid.UUID(normalized_headers["x-client-request-id"])
-
-    def test_marks_credentials_as_unredirected(self, mitm_ctx):
-        with (
-            mitm_ctx(),
-            patch.object(platform_api, "VERCEL_BYPASS", "secret-bypass-value"),
-        ):
-            req = platform_api.make_api_request(
-                "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-                b"{}",
-                "tok-xyz",
-            )
-
-        redirected_headers = {name.lower(): value for name, value in req.headers.items()}
-        unredirected_headers = {
-            name.lower(): value for name, value in req.unredirected_hdrs.items()
-        }
-        assert "authorization" not in redirected_headers
-        assert "x-vercel-protection-bypass" not in redirected_headers
-        assert unredirected_headers["authorization"] == "Bearer tok-xyz"
-        assert unredirected_headers["x-vercel-protection-bypass"] == "secret-bypass-value"
-
-    @pytest.mark.parametrize(
-        "url",
-        [
-            pytest.param("file:///etc/passwd", id="file"),
-            pytest.param("ftp://example.com/api", id="ftp"),
-            pytest.param(
-                "//api.vm0.ai/api/webhooks/agent/firewall/auth",
-                id="scheme-relative",
-            ),
-            pytest.param("https:path-without-host", id="https-without-host"),
-        ],
-    )
-    def test_rejects_non_absolute_http_urls(self, url: str):
-        with pytest.raises(ValueError, match="absolute http"):
-            platform_api.make_api_request(url, b"{}", "tok-xyz")
-
-
-class TestFetchFirewallHeaders:
-    async def test_sends_request_and_maps_basic_success(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            firewall_auth_success_response({"Authorization": "Bearer tok"})
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            result = await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    auth_headers={"Authorization": "Bearer ${{ secrets.TOKEN }}"}
-                ),
-            )
-
-        assert result.payload.headers == {"Authorization": "Bearer tok"}
-        assert result.payload.base is None
-        assert result.payload.query is None
-
-        assert endpoint.request_count == 1
-        request = endpoint.requests[0]
-        assert request.method == "POST"
-        assert request.path == "/api/webhooks/agent/firewall/auth"
-        assert request.headers["authorization"] == "Bearer tok-xyz"
-        assert request.headers["content-type"] == "application/json"
-        assert request.headers["user-agent"] == "vm0-mitm-addon/1.0"
-        assert request.headers["x-client-version"] == "runner-version-test"
-        assert request.headers["x-client-type"] == "MitmAddon"
-        assert request.headers["x-client-session-id"] == "runner-session-test"
-        uuid.UUID(request.headers["x-client-request-id"])
-        assert "x-vercel-protection-bypass" not in request.headers
-        assert request.json_body() == {
-            "encryptedSecrets": "iv:tag:data",
-            "authHeaders": {"Authorization": "Bearer ${{ secrets.TOKEN }}"},
-        }
-
-    async def test_success_response_shape_is_mapped(self, mitm_ctx):
-        expires_at = time.time() + 30
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            firewall_auth_success_response(
-                {
-                    "Authorization": "Bearer tok",
-                    "X-Custom": "custom",
-                },
-                expires_at=expires_at,
-                resolved_secrets=["API_TOKEN"],
-                refreshed_connectors=["notion"],
-                refreshed_secrets=["NOTION_TOKEN"],
-            )
-            | {
-                "base": "https://example.com/webhook/secret",
-                "query": {"api_key": "resolved-key"},
-                "futureField": {"ignored": True},
-            }
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            result = await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    auth_headers={
-                        "Authorization": "Bearer ${{ secrets.TOKEN }}",
-                        "X-Custom": "${{ secrets.CUSTOM }}",
-                    },
-                    auth_base="${{ secrets.WEBHOOK_URL }}",
-                    auth_query={"api_key": "${{ secrets.API_KEY }}"},
-                )
-            )
-
-        assert result.payload.headers == {
-            "Authorization": "Bearer tok",
-            "X-Custom": "custom",
-        }
-        assert result.payload.base == "https://example.com/webhook/secret"
-        assert result.payload.query == {"api_key": "resolved-key"}
-        assert result.payload.aws_sigv4 is None
-        assert result.expires_at == expires_at
-        assert result.payload.resolved_secrets == ["API_TOKEN"]
-        assert result.refreshed_connectors == ["notion"]
-        assert result.refreshed_secrets == ["NOTION_TOKEN"]
-        assert not hasattr(result, "futureField")
-
-    @pytest.mark.parametrize(
-        "session_token",
-        [
-            pytest.param(None, id="without-session-token"),
-            pytest.param("session-token", id="with-session-token"),
-        ],
-    )
-    async def test_sigv4_success_response_is_cached(
-        self,
-        mitm_ctx,
-        session_token: str | None,
-    ):
-        request_aws_sigv4 = {
-            "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-            "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-        }
-        response_aws_sigv4: dict[str, object] = {
-            "accessKeyId": "access-key-id",
-            "secretAccessKey": "secret-access-key",
-            "futureField": {"ignored": True},
-        }
-        if session_token is not None:
-            request_aws_sigv4["sessionToken"] = "${{ secrets.AWS_SESSION_TOKEN }}"
-            response_aws_sigv4["sessionToken"] = session_token
-
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            firewall_auth_success_response({})
-            | {
-                "awsSigv4": response_aws_sigv4,
-            }
-        )
-        cache_key = auth_cache_key()
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            result = await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(auth_aws_sigv4=request_aws_sigv4),
-            )
-
-        expected_credentials = AwsSigV4Credentials(
-            "access-key-id",
-            "secret-access-key",
-            session_token,
-        )
-        assert result["aws_sigv4"] == expected_credentials
-        assert require_cached_headers(cache_key).aws_sigv4 == expected_credentials
-
-    @pytest.mark.parametrize(
-        ("aws_sigv4", "expected_reason"),
-        [
-            pytest.param(None, "awsSigv4 must be an object", id="sigv4-null"),
-            pytest.param([], "awsSigv4 must be an object", id="sigv4-array"),
-            pytest.param(
-                {"secretAccessKey": "secret-access-key"},
-                "awsSigv4.accessKeyId is required",
-                id="access-key-missing",
-            ),
-            pytest.param(
-                {"accessKeyId": "", "secretAccessKey": "secret-access-key"},
-                "awsSigv4.accessKeyId is required",
-                id="access-key-empty",
-            ),
-            pytest.param(
-                {"accessKeyId": None, "secretAccessKey": "secret-access-key"},
-                "awsSigv4.accessKeyId is required",
-                id="access-key-null",
-            ),
-            pytest.param(
-                {"accessKeyId": 123, "secretAccessKey": "secret-access-key"},
-                "awsSigv4.accessKeyId is required",
-                id="access-key-number",
-            ),
-            pytest.param(
-                {"accessKeyId": "access-key-id"},
-                "awsSigv4.secretAccessKey is required",
-                id="secret-key-missing",
-            ),
-            pytest.param(
-                {"accessKeyId": "access-key-id", "secretAccessKey": ""},
-                "awsSigv4.secretAccessKey is required",
-                id="secret-key-empty",
-            ),
-            pytest.param(
-                {"accessKeyId": "access-key-id", "secretAccessKey": None},
-                "awsSigv4.secretAccessKey is required",
-                id="secret-key-null",
-            ),
-            pytest.param(
-                {"accessKeyId": "access-key-id", "secretAccessKey": 123},
-                "awsSigv4.secretAccessKey is required",
-                id="secret-key-number",
-            ),
-            pytest.param(
-                {
-                    "accessKeyId": "access-key-id",
-                    "secretAccessKey": "secret-access-key",
-                    "sessionToken": "",
-                },
-                "sessionToken must not be empty",
-                id="session-token-empty",
-            ),
-            pytest.param(
-                {
-                    "accessKeyId": "access-key-id",
-                    "secretAccessKey": "secret-access-key",
-                    "sessionToken": None,
-                },
-                "sessionToken must be a string",
-                id="session-token-null",
-            ),
-            pytest.param(
-                {
-                    "accessKeyId": "access-key-id",
-                    "secretAccessKey": "secret-access-key",
-                    "sessionToken": 123,
-                },
-                "sessionToken must be a string",
-                id="session-token-number",
-            ),
-        ],
-    )
-    async def test_malformed_sigv4_response_is_not_cached(
-        self,
-        mitm_ctx,
-        aws_sigv4: object,
-        expected_reason: str,
-    ):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(firewall_auth_success_response({}) | {"awsSigv4": aws_sigv4})
-        cache_key = auth_cache_key()
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(ValueError, match=_MALFORMED_SUCCESS_PREFIX) as exc_info,
-        ):
-            await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(
-                    auth_aws_sigv4={
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                    }
-                ),
-            )
-
-        assert str(exc_info.value) == f"{_MALFORMED_SUCCESS_PREFIX}: {expected_reason}"
-        assert cached_headers(cache_key) is None
-
-    @pytest.mark.parametrize(
-        ("auth_request", "response"),
-        [
-            (
-                _firewall_auth_request(auth_base="${{ secrets.WEBHOOK_URL }}"),
-                firewall_auth_success_response({}),
-            ),
-            (
-                _firewall_auth_request(),
-                firewall_auth_success_response({}) | {"base": "https://hooks.example.com/secret"},
-            ),
-            (
-                _firewall_auth_request(auth_base="${{ secrets.WEBHOOK_URL }}"),
-                firewall_auth_success_response({}) | {"base": ""},
-            ),
-            (
-                _firewall_auth_request(
-                    auth_aws_sigv4={
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                    }
-                ),
-                firewall_auth_success_response({}),
-            ),
-            (
-                _firewall_auth_request(),
-                firewall_auth_success_response({})
-                | {
-                    "awsSigv4": {
-                        "accessKeyId": "access-key-id",
-                        "secretAccessKey": "secret-access-key",
-                    },
-                },
-            ),
-            (
-                _firewall_auth_request(
-                    auth_aws_sigv4={
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                    }
-                ),
-                firewall_auth_success_response({})
-                | {
-                    "awsSigv4": {
-                        "accessKeyId": "access-key-id",
-                        "secretAccessKey": "secret-access-key",
-                        "sessionToken": "session-token",
-                    },
-                },
-            ),
-            (
-                _firewall_auth_request(
-                    auth_aws_sigv4={
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                        "sessionToken": "${{ secrets.AWS_SESSION_TOKEN }}",
-                    }
-                ),
-                firewall_auth_success_response({})
-                | {
-                    "awsSigv4": {
-                        "accessKeyId": "access-key-id",
-                        "secretAccessKey": "secret-access-key",
-                    },
-                },
-            ),
-            (
-                _firewall_auth_request(auth_headers={"Authorization": "template"}),
-                firewall_auth_success_response({"X-Unexpected": "value"}),
-            ),
-            (
-                _firewall_auth_request(auth_query={"api_key": "template"}),
-                firewall_auth_success_response({}) | {"query": {"unexpected": "value"}},
-            ),
-        ],
-        ids=[
-            "missing-base",
-            "unexpected-base",
-            "empty-base",
-            "missing-sigv4",
-            "unexpected-sigv4",
-            "unexpected-session-token",
-            "missing-session-token",
-            "header-name-mismatch",
-            "query-name-mismatch",
-        ],
-    )
-    async def test_rejects_response_inconsistent_with_request(
-        self,
-        mitm_ctx,
-        auth_request: auth_client.FirewallAuthRequest,
-        response: dict[str, object],
-    ):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(response)
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(ValueError, match=_MALFORMED_SUCCESS_PREFIX),
-        ):
-            await auth_client.fetch_firewall_headers(auth_request)
-
-    async def test_inconsistent_response_is_not_cached(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(firewall_auth_success_response({}))
-        cache_key = auth_cache_key()
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(ValueError, match=_MALFORMED_SUCCESS_PREFIX),
-        ):
-            await auth_cache.get_firewall_headers(
-                cache_key,
-                _firewall_auth_request(auth_base="${{ secrets.WEBHOOK_URL }}"),
-            )
-
-        assert cached_headers(cache_key) is None
-
-    async def test_sends_optional_request_body_fields(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            firewall_auth_success_response({}, expires_at=time.time() + 30)
-            | {
-                "base": "https://hooks.example.com/secret",
-                "query": {"api_key": "resolved-key"},
-            }
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    secret_connector_map={"TOKEN": "notion"},
-                    secret_connector_metadata_map={"TOKEN": {"kind": "oauth"}},
-                    vars_map={"TEAM": "vm0"},
-                    auth_base="${{ secrets.WEBHOOK_URL }}",
-                    auth_query={"api_key": "${{ secrets.API_KEY }}"},
-                    firewall_billable=True,
-                ),
-                force_refresh=True,
-            )
-
-        body = endpoint.requests[0].json_body()
-        assert body["encryptedSecrets"] == "iv:tag:data"
-        assert body["authHeaders"] == {}
-        assert body["secretConnectorMap"] == {"TOKEN": "notion"}
-        assert body["secretConnectorMetadataMap"] == {"TOKEN": {"kind": "oauth"}}
-        assert body["vars"] == {"TEAM": "vm0"}
-        assert body["authBase"] == "${{ secrets.WEBHOOK_URL }}"
-        assert body["authQuery"] == {"api_key": "${{ secrets.API_KEY }}"}
-        assert "authAwsSigv4" not in body
-        assert body["firewallBillable"] is True
-        assert body["forceRefresh"] is True
-        assert "firewallName" not in body
-        assert "modelUsageProvider" not in body
-
-    async def test_sends_sigv4_request_body(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            firewall_auth_success_response({})
-            | {
-                "awsSigv4": {
-                    "accessKeyId": "access-key-id",
-                    "secretAccessKey": "secret-access-key",
-                    "sessionToken": "session-token",
-                },
-            }
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    auth_aws_sigv4={
-                        "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                        "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                        "sessionToken": "${{ secrets.AWS_SESSION_TOKEN }}",
-                    }
-                ),
-            )
-
-        assert endpoint.requests[0].json_body() == {
-            "encryptedSecrets": "iv:tag:data",
-            "authHeaders": {},
-            "authAwsSigv4": {
-                "accessKeyId": "${{ secrets.AWS_ACCESS_KEY_ID }}",
-                "secretAccessKey": "${{ secrets.AWS_SECRET_ACCESS_KEY }}",
-                "sessionToken": "${{ secrets.AWS_SESSION_TOKEN }}",
-            },
-        }
-
-    async def test_omits_empty_optional_request_body_fields(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(firewall_auth_success_response({}))
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    auth_query={},
-                    secret_connector_map={},
-                    secret_connector_metadata_map={},
-                    vars_map={},
-                    firewall_billable=False,
-                ),
-                force_refresh=False,
-            )
-
-        assert endpoint.requests[0].json_body() == {
-            "encryptedSecrets": "iv:tag:data",
-            "authHeaders": {},
-        }
-
-    def test_request_repr_omits_sensitive_values(self):
-        request = _firewall_auth_request(
-            encrypted_secrets="secret-encrypted-payload",
-            sandbox_auth="secret-sandbox-token",
-        )
-
-        rendered = repr(request)
-
-        assert "secret-encrypted-payload" not in rendered
-        assert "secret-sandbox-token" not in rendered
-
-    async def test_includes_vercel_bypass_header(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(firewall_auth_success_response({}))
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", "secret-bypass-value"),
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert endpoint.requests[0].headers["x-vercel-protection-bypass"] == "secret-bypass-value"
-
-    @pytest.mark.parametrize("status", [301, 302, 303])
-    async def test_rejects_cross_origin_redirect_without_forwarding_credentials(
-        self,
-        status: int,
-        mitm_ctx,
-    ):
-        source = FakeAuthEndpoint()
-        target = FakeAuthEndpoint()
-
-        with target.run():
-            source.queue_response(
-                status,
-                headers=(("Location", f"{target.api_url}/redirected"),),
-            )
-            with (
-                source.run(),
-                mitm_ctx(api_url=source.api_url),
-                patch.object(platform_api, "VERCEL_BYPASS", "secret-bypass-value"),
-                pytest.raises(urllib.error.HTTPError) as exc_info,
-            ):
-                await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert exc_info.value.code == status
-        assert source.request_count == 1
-        request = source.requests[0]
-        assert request.method == "POST"
-        assert request.headers["authorization"] == "Bearer tok-xyz"
-        assert request.headers["x-vercel-protection-bypass"] == "secret-bypass-value"
-        assert target.requests == ()
-
-    async def test_invalid_api_url_raises_before_open(self):
-        with (
-            patch.object(platform_api, "get_api_url", return_value="file:///etc/passwd"),
-            patch("firewall_auth_client._opener.open") as mock_open,
-            pytest.raises(ValueError, match="absolute http"),
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        mock_open.assert_not_called()
-
-    async def test_424_connector_not_configured_raises_custom_error(self, mitm_ctx):
-        """Auth endpoint 424 CONNECTOR_NOT_CONFIGURED raises ConnectorNotConfiguredError."""
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            {
-                "error": {
-                    "message": "Connector not configured",
-                    "code": "CONNECTOR_NOT_CONFIGURED",
-                }
-            },
-            status=424,
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            with pytest.raises(auth_client.ConnectorNotConfiguredError) as exc_info:
-                await auth_client.fetch_firewall_headers(_firewall_auth_request())
-            assert "Connector not configured" in str(exc_info.value)
-
-    async def test_402_insufficient_credits_raises_custom_error(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            {
-                "error": {
-                    "message": "Insufficient credits",
-                    "code": "INSUFFICIENT_CREDITS",
-                }
-            },
-            status=402,
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            with pytest.raises(auth_client.InsufficientCreditsError) as exc_info:
-                await auth_client.fetch_firewall_headers(_firewall_auth_request())
-            assert "Insufficient credits" in str(exc_info.value)
-
-    @pytest.mark.parametrize(
-        (
-            "status",
-            "code",
-            "message",
-            "connectors",
-            "failure_reason",
-            "expected_failure_reason",
-        ),
-        [
-            (
-                424,
-                "TOKEN_ACCESS_RESOLUTION_FAILED",
-                "Token access resolution failed for: notion.",
-                ["notion"],
-                None,
-                None,
-            ),
-            (
-                403,
-                "FORBIDDEN",
-                "Invalid model-provider secret owner",
-                None,
-                None,
-                None,
-            ),
-            (
-                502,
-                "TOKEN_REFRESH_FAILED",
-                "Access token expired and refresh failed for: codex-oauth-token.",
-                ["codex-oauth-token"],
-                "upstream_provider",
-                "upstream_provider",
-            ),
-            (
-                502,
-                "TOKEN_REFRESH_FAILED",
-                "Access token expired and refresh failed for: notion.",
-                ["notion"],
-                "provider_rate_limited",
-                None,
-            ),
-        ],
-        ids=[
-            "token-access-resolution",
-            "forbidden",
-            "token-refresh",
-            "unknown-failure-reason",
-        ],
-    )
-    async def test_current_structured_error_raises_custom_error(
-        self,
-        mitm_ctx,
-        status: int,
-        code: str,
-        message: str,
-        connectors: list[str] | None,
-        failure_reason: str | None,
-        expected_failure_reason: str | None,
-    ):
-        """Current auth endpoint errors should preserve their code and connectors."""
-        error_info: dict[str, object] = {
-            "message": message,
-            "code": code,
-        }
-        if connectors is not None:
-            error_info["connectors"] = connectors
-        if failure_reason is not None:
-            error_info["failureReason"] = failure_reason
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response({"error": error_info}, status=status)
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(auth_client.FirewallAuthApiError) as exc_info,
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert exc_info.value.status == status
-        assert exc_info.value.code == code
-        assert str(exc_info.value) == message
-        assert exc_info.value.connectors == connectors
-        assert exc_info.value.failure_reason == expected_failure_reason
-
-    async def test_structured_http_error_at_body_limit_is_preserved(self, mitm_ctx):
-        error_body = json.dumps(
-            {
-                "error": {
-                    "message": "Access token expired and refresh failed for: notion.",
-                    "code": "TOKEN_REFRESH_FAILED",
-                }
-            }
-        ).encode()
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_response(502, body=error_body)
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(
-                auth_client,
-                "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES",
-                len(error_body),
-            ),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(auth_client.FirewallAuthApiError) as exc_info,
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert exc_info.value.code == "TOKEN_REFRESH_FAILED"
-
-    async def test_http_error_over_body_limit_raises(self, mitm_ctx):
-        error_body = json.dumps(
-            {
-                "error": {
-                    "message": "Access token expired and refresh failed for: notion.",
-                    "code": "TOKEN_REFRESH_FAILED",
-                }
-            }
-        ).encode()
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_response(502, body=error_body)
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(
-                auth_client,
-                "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES",
-                len(error_body) - 1,
-            ),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(
-                auth_client.FirewallAuthResponseTooLargeError,
-                match="Firewall auth response body too large",
-            ),
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-    @pytest.mark.parametrize(
-        "error_body",
-        [
-            pytest.param(b"\xff", id="invalid-utf8"),
-            b"not-json",
-            b'"plain string"',
-            b"[1, 2, 3]",
-            b"{}",
-            json.dumps({"error": "not-a-dict"}).encode(),
-            json.dumps({"error": None}).encode(),
-            json.dumps({"error": {}}).encode(),
-            json.dumps({"error": {"message": "Bad Request", "code": "BAD_REQUEST"}}).encode(),
-        ],
-    )
-    async def test_malformed_http_error_envelope_reraises_http_error(
-        self,
-        mitm_ctx,
-        error_body: bytes,
-    ):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_response(400, body=error_body)
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(urllib.error.HTTPError) as exc_info,
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert exc_info.value.code == 400
-
-    @pytest.mark.parametrize(
-        ("code", "status", "exception_type", "default_message"),
-        [
-            (
-                "CONNECTOR_NOT_CONFIGURED",
-                424,
-                auth_client.ConnectorNotConfiguredError,
-                "Connector not configured",
-            ),
-            (
-                "INSUFFICIENT_CREDITS",
-                402,
-                auth_client.InsufficientCreditsError,
-                "Insufficient credits",
-            ),
-        ],
-    )
-    async def test_known_error_with_non_string_message_uses_default(
-        self,
-        mitm_ctx,
-        code: str,
-        status: int,
-        exception_type: type[Exception],
-        default_message: str,
-    ):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(
-            {
-                "error": {
-                    "message": None,
-                    "code": code,
-                }
-            },
-            status=status,
-        )
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(exception_type) as exc_info,
-        ):
-            await auth_client.fetch_firewall_headers(_firewall_auth_request())
-
-        assert str(exc_info.value) == default_message
-
-    async def test_async_wrapper_uses_api_url_from_ctx(self, mitm_ctx):
-        endpoint = FakeAuthEndpoint()
-        endpoint.queue_json_response(firewall_auth_success_response({"Auth": "tok"}))
-
-        with (
-            endpoint.run(),
-            mitm_ctx(api_url=endpoint.api_url),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            result = await auth_client.fetch_firewall_headers(
-                _firewall_auth_request(
-                    encrypted_secrets="enc",
-                    auth_headers={"Auth": "${{ secrets.TOKEN }}"},
-                    sandbox_auth="sandbox-tok",
-                )
-            )
-
-        assert result.payload.headers == {"Auth": "tok"}
-        assert endpoint.requests[0].path == "/api/webhooks/agent/firewall/auth"
-
-
-class TestFirewallAuthSuccessParser:
-    @pytest.mark.parametrize(
-        "body",
-        [
-            pytest.param([], id="array"),
-            pytest.param(None, id="null"),
-            pytest.param("plain string", id="string"),
-            pytest.param(123, id="number"),
-            pytest.param(
-                {
-                    "expiresAt": None,
-                    "resolvedSecrets": [],
-                    "refreshedConnectors": [],
-                    "refreshedSecrets": [],
-                },
-                id="missing-headers",
-            ),
-            pytest.param(firewall_auth_success_response({}) | {"headers": []}, id="headers-array"),
-            pytest.param(
-                firewall_auth_success_response({}) | {"headers": {"Authorization": 123}},
-                id="header-value-number",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"base": []},
-                id="base-array",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"base": ""},
-                id="base-empty",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"query": []},
-                id="query-array",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"query": {"api_key": 123}},
-                id="query-value-number",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"resolvedSecrets": "TOKEN"},
-                id="resolved-secrets-string",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"refreshedConnectors": [123]},
-                id="refreshed-connectors-number",
-            ),
-            pytest.param(
-                firewall_auth_success_response({}) | {"refreshedSecrets": [None]},
-                id="refreshed-secrets-null",
-            ),
-        ],
-    )
-    def test_malformed_success_response_shape_raises_value_error(self, body: object):
-        with pytest.raises(ValueError, match=_MALFORMED_SUCCESS_PREFIX):
-            auth_client._parse_firewall_auth_success(body, _firewall_auth_request())
-
-
-class TestFirewallAuthResponseBodyReader:
-    def test_response_at_body_limit_is_accepted(self):
-        response_body = json.dumps({"headers": {}}).encode()
-        mock_resp = _raw_response(response_body)
-
-        with patch.object(auth_client, "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES", len(response_body)):
-            assert auth_client._read_firewall_auth_response_body(mock_resp) == response_body
-
-        mock_resp.read.assert_called_once_with(len(response_body) + 1)
-
-    def test_response_over_body_limit_raises(self):
-        response_body = json.dumps({"headers": {}}).encode()
-        mock_resp = _raw_response(response_body)
-
-        with (
-            patch.object(
-                auth_client, "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES", len(response_body) - 1
-            ),
-            pytest.raises(
-                auth_client.FirewallAuthResponseTooLargeError,
-                match="Firewall auth response body too large",
-            ),
-        ):
-            auth_client._read_firewall_auth_response_body(mock_resp)
-
-        mock_resp.read.assert_called_once_with(len(response_body))
-
-
-class TestFetchFirewallHeadersResourceBoundary:
-    def test_closes_response_on_success(self, mitm_ctx):
-        """Success path must close the HTTP response — FD leak guard (#10475)."""
-        mock_resp = MagicMock()
-        mock_resp.__enter__.return_value = mock_resp
-        mock_resp.read.return_value = json.dumps(firewall_auth_success_response({})).encode()
-
-        with (
-            mitm_ctx(),
-            patch("platform_api.urllib.request.Request"),
-            patch("firewall_auth_client._opener.open", return_value=mock_resp),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-        ):
-            auth_client._fetch_firewall_headers_sync(_firewall_auth_request(), "https://api.vm0.ai")
-
-        mock_resp.__exit__.assert_called_once()  # urllib external boundary (#9991)
-
-    def test_closes_http_error_response_when_body_is_unreadable(self, mitm_ctx):
-        http_error = urllib.error.HTTPError(
-            "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-            400,
-            "Bad Request",
-            Message(),
-            _UnreadableHttpErrorBody(),
-        )
-        http_error.close = MagicMock()
-
-        with (
-            mitm_ctx(),
-            patch("platform_api.urllib.request.Request"),
-            patch("firewall_auth_client._opener.open", side_effect=http_error),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(urllib.error.HTTPError) as exc_info,
-        ):
-            auth_client._fetch_firewall_headers_sync(_firewall_auth_request(), "https://api.vm0.ai")
-
-        assert exc_info.value is http_error
-        http_error.close.assert_called_once()
-
-    def test_closes_http_error_response_when_body_has_invalid_utf8(self, mitm_ctx):
-        http_error = _http_error(
-            "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-            400,
-            "Bad Request",
-            b"\xff",
-        )
-        http_error.close = MagicMock()
-
-        with (
-            mitm_ctx(),
-            patch("platform_api.urllib.request.Request"),
-            patch("firewall_auth_client._opener.open", side_effect=http_error),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(urllib.error.HTTPError) as exc_info,
-        ):
-            auth_client._fetch_firewall_headers_sync(_firewall_auth_request(), "https://api.vm0.ai")
-
-        assert exc_info.value is http_error
-        assert exc_info.value.code == 400
-        http_error.close.assert_called_once()
-
-    def test_closes_http_error_response_when_body_is_too_large(self, mitm_ctx):
-        error_body = json.dumps(
-            {
-                "error": {
-                    "message": "Access token expired and refresh failed for: notion.",
-                    "code": "TOKEN_REFRESH_FAILED",
-                }
-            }
-        ).encode()
-        http_error = _http_error(
-            "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-            502,
-            "Bad Gateway",
-            error_body,
-        )
-        http_error.close = MagicMock()
-
-        with (
-            mitm_ctx(),
-            patch.object(
-                auth_client,
-                "MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES",
-                len(error_body) - 1,
-            ),
-            patch("platform_api.urllib.request.Request"),
-            patch("firewall_auth_client._opener.open", side_effect=http_error),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(
-                auth_client.FirewallAuthResponseTooLargeError,
-                match="Firewall auth response body too large",
-            ),
-        ):
-            auth_client._fetch_firewall_headers_sync(_firewall_auth_request(), "https://api.vm0.ai")
-
-        http_error.close.assert_called_once()
-
-    @pytest.mark.parametrize(
-        ("error_body", "expected_exception"),
-        [
-            (
-                json.dumps(
-                    {
-                        "error": {
-                            "message": "Access token expired and refresh failed for: notion.",
-                            "code": "TOKEN_REFRESH_FAILED",
-                        }
-                    }
-                ).encode(),
-                auth_client.FirewallAuthApiError,
-            ),
-            (b"{}", urllib.error.HTTPError),
-        ],
-    )
-    def test_closes_http_error_response(
-        self,
-        mitm_ctx,
-        error_body: bytes,
-        expected_exception: type[Exception],
-    ):
-        """HTTPError path must close the underlying socket — FD leak guard (#10475)."""
-        http_error = _http_error(
-            "https://api.vm0.ai/api/webhooks/agent/firewall/auth",
-            400,
-            "Bad Request",
-            error_body,
-        )
-        http_error.close = MagicMock()
-
-        with (
-            mitm_ctx(),
-            patch("platform_api.urllib.request.Request"),
-            patch("firewall_auth_client._opener.open", side_effect=http_error),
-            patch.object(platform_api, "VERCEL_BYPASS", ""),
-            pytest.raises(expected_exception),
-        ):
-            auth_client._fetch_firewall_headers_sync(_firewall_auth_request(), "https://api.vm0.ai")
-
-        http_error.close.assert_called_once()  # urllib external boundary (#9991)
